Make kf:commit a blocking function
kf:commit was previously an async function which would return a
future. This ended up being a bad API choice because most usages of
kf:commit were intended to be blocking. For example, everytime I used
kf:commit, I'd block immediately with kf:value:

  (kf:value (kf:commit consumer))

So kf:commit will now block and either return nil on success or a
kf:kafka-error on failure.

This was also the only usage of kf:future so high-level/future.lisp
was deleted.